Make Navbar and Login page global
